requirements.txt
streamlit==1.33.0
pandas==1.5.3
SQLAlchemy==2.0.9
psycopg2-binary==2.9.6
plotly==5.14.1
python-dotenv==1.0.0
Pillow==9.5.0
pydeck==0.8.0
streamlit-aggrid==0.3.4
bcrypt==4.0.1
numpy==1.23.5
extra-streamlit-components==0.1.2
streamlit-extras
google-generativeai
openai